+++ tgui/packages/tgui/interfaces/NtosRCON.tsx
@@ -1,5 +1,5 @@
 import { NtosWindow } from '../layouts';
-import { RCONContent } from './RCON';
+import { RCONContent } from './RCON/RCONContent';
 
 export const NtosRCON = () => {
   return (
