+++ tgui/packages/tgui/interfaces/PowerMonitor/PowerMonitorFocus.tsx
@@ -3,113 +3,57 @@ import { flow } from 'common/fp';
 import { toFixed } from 'common/math';
 import { useState } from 'react';
 
-import { useBackend } from '../backend';
+import { useBackend } from '../../backend';
 import {
   Box,
   Button,
   Chart,
-  ColorBox,
   Flex,
-  Icon,
   LabeledList,
   ProgressBar,
   Section,
   Table,
-} from '../components';
-import { Window } from '../layouts';
+} from '../../components';
+import { PEAK_DRAW } from './constants';
+import { powerRank } from './functions';
+import { AreaCharge, AreaStatusColorBox } from './PowerMonitorHelpers';
+import { area, sensor } from './types';
 
-const PEAK_DRAW = 500000;
-
-export const powerRank = (str) => {
-  const unit = String(str.split(' ')[1]).toLowerCase();
-  return ['w', 'kw', 'mw', 'gw'].indexOf(unit);
-};
-
-export const PowerMonitor = () => {
-  return (
-    <Window width={550} height={700}>
-      <Window.Content scrollable>
-        <PowerMonitorContent />
-      </Window.Content>
-    </Window>
-  );
-};
-
-export const PowerMonitorContent = (props) => {
-  const { act, data } = useBackend();
-
-  const { map_levels, all_sensors, focus } = data;
-
-  if (focus) {
-    return <PowerMonitorFocus focus={focus} />;
-  }
-
-  let body = <Box color="bad">No sensors detected</Box>;
-
-  if (all_sensors) {
-    body = (
-      <Table>
-        {all_sensors.map((sensor) => (
-          <Table.Row key={sensor.name}>
-            <Table.Cell>
-              <Button
-                icon={sensor.alarm ? 'bell' : 'sign-in-alt'}
-                onClick={() => act('setsensor', { id: sensor.name })}
-              >
-                {sensor.name}
-              </Button>
-            </Table.Cell>
-          </Table.Row>
-        ))}
-      </Table>
-    );
-  }
-
-  return (
-    <Section
-      title="No active sensor. Listing all."
-      buttons={
-        <Button icon="undo" onClick={() => act('refresh')}>
-          Scan For Sensors
-        </Button>
-      }
-    >
-      {body}
-    </Section>
-  );
-};
-
-export const PowerMonitorFocus = (props) => {
-  const { act, data } = useBackend();
+export const PowerMonitorFocus = (props: { focus: sensor }) => {
+  const { act } = useBackend();
   const { focus } = props;
   const { history } = focus;
-  const [sortByField, setSortByField] = useState(null);
-  const supply = history.supply[history.supply.length - 1] || 0;
-  const demand = history.demand[history.demand.length - 1] || 0;
-  const supplyData = history.supply.map((value, i) => [i, value]);
-  const demandData = history.demand.map((value, i) => [i, value]);
-  const maxValue = Math.max(PEAK_DRAW, ...history.supply, ...history.demand);
+  const [sortByField, setSortByField] = useState<string | null>(null);
+  const supply: number = history.supply[history.supply.length - 1] || 0;
+  const demand: number = history.demand[history.demand.length - 1] || 0;
+  const supplyData: number[][] = history.supply.map((value, i) => [i, value]);
+  const demandData: number[][] = history.demand.map((value, i) => [i, value]);
+  const maxValue: number = Math.max(
+    PEAK_DRAW,
+    ...history.supply,
+    ...history.demand,
+  );
   // Process area data
-  const areas = flow([
-    map((area, i) => ({
+  const areas: area[] = flow([
+    map((area: area, i) => ({
       ...area,
       // Generate a unique id
       id: area.name + i,
     })),
-    sortByField === 'name' && sortBy((area) => area.name),
-    sortByField === 'charge' && sortBy((area) => -area.charge),
+    sortByField === 'name' && sortBy((area: area) => area.name),
+    sortByField === 'charge' && sortBy((area: area) => -area.charge),
     sortByField === 'draw' &&
       sortBy(
-        (area) => -powerRank(area.load),
-        (area) => -parseFloat(area.load),
+        (area: area) => -powerRank(area.load),
+        (area: area) => -parseFloat(area.load),
       ),
     sortByField === 'problems' &&
       sortBy(
-        (area) => area.eqp,
-        (area) => area.lgt,
-        (area) => area.env,
-        (area) => area.charge,
-        (area) => area.name,
+        (area: area) => area.eqp,
+        (area: area) => area.lgt,
+        (area: area) => area.env,
+        (area: area) => area.charge,
+        (area: area) => area.name,
       ),
   ])(focus.areas);
   return (
@@ -177,26 +121,32 @@ export const PowerMonitorFocus = (props) => {
           </Box>
           <Button.Checkbox
             checked={sortByField === 'name'}
-            onClick={() => setSortByField(sortByField !== 'name' && 'name')}
+            onClick={() =>
+              setSortByField((sortByField !== 'name' && 'name') || null)
+            }
           >
             Name
           </Button.Checkbox>
           <Button.Checkbox
             checked={sortByField === 'charge'}
-            onClick={() => setSortByField(sortByField !== 'charge' && 'charge')}
+            onClick={() =>
+              setSortByField((sortByField !== 'charge' && 'charge') || null)
+            }
           >
             Charge
           </Button.Checkbox>
           <Button.Checkbox
             checked={sortByField === 'draw'}
-            onClick={() => setSortByField(sortByField !== 'draw' && 'draw')}
+            onClick={() =>
+              setSortByField((sortByField !== 'draw' && 'draw') || null)
+            }
           >
             Draw
           </Button.Checkbox>
           <Button.Checkbox
             checked={sortByField === 'problems'}
             onClick={() =>
-              setSortByField(sortByField !== 'problems' && 'problems')
+              setSortByField((sortByField !== 'problems' && 'problems') || null)
             }
           >
             Problems
@@ -218,7 +168,7 @@ export const PowerMonitorFocus = (props) => {
             </Table.Cell>
           </Table.Row>
           {areas.map((area, i) => (
-            <tr key={area.id} className="Table__row candystripe">
+            <tr key={i} className="Table__row candystripe">
               <td>{area.name}</td>
               <td className="Table__cell text-right text-nowrap">
                 <AreaCharge charging={area.charging} charge={area.charge} />
@@ -240,45 +190,5 @@ export const PowerMonitorFocus = (props) => {
         </Table>
       </Section>
     </>
-  );
-};
-
-export const AreaCharge = (props) => {
-  const { charging, charge } = props;
-  return (
-    <>
-      <Icon
-        width="18px"
-        textAlign="center"
-        name={
-          (charging === 0 &&
-            (charge > 50 ? 'battery-half' : 'battery-quarter')) ||
-          (charging === 1 && 'bolt') ||
-          (charging === 2 && 'battery-full')
-        }
-        color={
-          (charging === 0 && (charge > 50 ? 'yellow' : 'red')) ||
-          (charging === 1 && 'yellow') ||
-          (charging === 2 && 'green')
-        }
-      />
-      <Box inline width="36px" textAlign="right">
-        {toFixed(charge) + '%'}
-      </Box>
-    </>
-  );
-};
-
-const AreaStatusColorBox = (props) => {
-  const { status } = props;
-  const power = Boolean(status & 2);
-  const mode = Boolean(status & 1);
-  const tooltipText = (power ? 'On' : 'Off') + ` [${mode ? 'auto' : 'manual'}]`;
-  return (
-    <ColorBox
-      color={power ? 'good' : 'bad'}
-      content={mode ? undefined : 'M'}
-      title={tooltipText}
-    />
   );
 };
