+++ tgui/packages/tgui/interfaces/PowerMonitor/constants.ts
@@ -1,0 +1,1 @@
+export const PEAK_DRAW: number = 500000;
