+++ tgui/packages/tgui/interfaces/RCON/constants.ts
@@ -1,0 +1,2 @@
+// Common power multiplier
+export const POWER_MUL: number = 1e3;
