+++ tgui/packages/tgui/interfaces/RequestConsole/constants.ts
@@ -1,0 +1,9 @@
+export const RCS_MAINMENU = 0; // Settings menu
+export const RCS_RQASSIST = 1; // Request supplies
+export const RCS_RQSUPPLY = 2; // Request assistance
+export const RCS_SENDINFO = 3; // Relay information
+export const RCS_SENTPASS = 4; // Message sent successfully
+export const RCS_SENTFAIL = 5; // Message sent unsuccessfully
+export const RCS_VIEWMSGS = 6; // View messages
+export const RCS_MESSAUTH = 7; // Authentication before sending
+export const RCS_ANNOUNCE = 8; // Send announcement
