+++ tgui/packages/tgui/interfaces/ResleevingConsole/constants.ts
@@ -1,0 +1,3 @@
+export const MENU_MAIN = 1;
+export const MENU_BODY = 2;
+export const MENU_MIND = 3;
